Feat: 닫기 버튼 클릭 시 폼 초기화
