+++ src/styles/layouts/_user.scss
@@ -16,12 +16,17 @@
         padding: 1rem;
         background-color: #fff;
 
+        &.detail {
+          
+        }
+
         .user__image {
           width: 50px;
           height: 50px;
           background-color: $bg-gray-400;
           margin-right: 0.5rem;
           border-radius: 50%;
+          object-fit: cover;
         }
 
         .user__info {
