+++ src/upload.js
@@ -1,6 +1,8 @@
 import { uploadImageToStorage, uploadInfoToDatabase } from "./uploadImage";
 
 const addForm = document.querySelector(".form-add");
+const imageInput = document.querySelector(".form__input-image");
+const imagePreview = document.querySelector(".form__image-box");
 
 addForm.addEventListener("submit", async (e) => {
   console.log("form submitted");
@@ -8,7 +10,7 @@ addForm.addEventListener("submit", async (e) => {
 
   const name = document.getElementById("name").value;
   const email = document.getElementById("email").value;
-  const rank = document.getElementById("rank").value;
+  const position = document.getElementById("position").value;
   const imageFile = document.getElementById("chooseFile").files[0];
 
   const imageUrl = await uploadImageToStorage(imageFile);
@@ -16,13 +18,14 @@ addForm.addEventListener("submit", async (e) => {
   const userData = {
     name,
     email,
-    rank,
+    position,
     imageUrl,
   };
 
   uploadInfoToDatabase(userData);
 
   addForm.reset();
-  document.querySelector(".form__image-box").classList.add("hidden");
-  document.querySelector(".form__input-image").classList.remove("hidden");
+  addForm.classList.add("hidden");
+  imageInput.classList.remove("invisible");
+  imagePreview.classList.add("hidden");
 });
